Helm: Decouple the Canary from self-monitoring
The Canary has no dependency on the Grafana Agent-based self-monitoring.

Only the automated test requires this. So we also disable the validation
for `Loki Canary requires self monitoring to also be enabled`.

This changes the path to the canary values from

    monitoring.selfMonitoring.lokiCanary

to

    monitoring.lokiCanary

Which is a breaking change to the values. But otherwise the logic change
is not clear enough.

Fixes #7364