loki: Make common.compactor_address dynamic
The setting `common.compactor_address` had a static value of `http://127.0.0.1`.
Let it depend on `server.http_listen_address` and `server.http_tls_config`.

While running all-in-one Loki instance, it's not possible to listen on anything
other than IPv4. This enables listening on only IPv6.

Since the default value for `http_listen_address` is `localhost` and the
default value for `compactor_address` is `127.0.0.1`, this actually modifies
the current behavior, not much though!

Signed-off-by: Josef Johansson <josef@oderland.se>