Loki: Add new sharding metrics (histogram based) (#7762)
**What this PR does / why we need it**:
Add histogram-version of two existing stream sharding metrics. The new
metrics are:
- `rate_store_stream_shards`: Will show us the distribution of shards
reported by ingesters. This will help us on monitoring the number of
shards we're dealing with. As of now we only have access to the max
number of shards seen, which is helpful but not enough.
- `rate_store_stream_rate_bytes`: Same as below: will help us understand
the distribution of stream sizes instead of only showing us the max
stream size seen.