Remove hardwired '127.0.0.1' values in default addresses (#7784)
While running all-in-one Loki instance, it's not possible to listen on anything other than IPv4 since compactor_address is then set to http://127.0.0.1. This enables listening on only IPv6.

Also if frontend_address, scheduler_ring and scheduler_address is not set, the default value is to connect to '127.0.0.1:<grpc_listen_port>'. Which means that grpc_listen_address is ignored.

**What this PR does / why we need it**:

It's not possible to use compactor while using an all-in-one instance and listening to IPv6.
Let's not ignore grpc_listen_address to not confuse users.

**Special notes for your reviewer**:

As noted in the first commit, the default behavior is changed from defaulting to `127.0.0.1` to `localhost`. Not sure how major that change is though. My /etc/hosts says
```
$ grep localhost /etc/hosts
127.0.0.1 localhost
```

Signed-off-by: Josef Johansson <josef@oderland.se>