Fix alertmanager per tenant override config behavior (#7805)
**Context:** #7359 implement multi tenancy for the alertmanager config.
However, the behavior of the per-tenant alertmanager config is
erroneous, as it tries to merge the global alertmanager config with the
per-tenant config. This does not work because there is no way of knowing
if certain values were set by the user or if they were undefined and
have the default Golang zero value (typically booleans).

This PR changes the behavior of the alertmanager config to override all
the configuration instead of trying to merge the global config with the
per-tenant config. It also applies default values when appropriate.

(cherry picked from commit 03a34db8eb248ec8a7261c7c889ee33260162208)