Promtail: Fix broken promtail-debug docker image (#7852)
**What this PR does / why we need it**:
The 'grafana/loki-build-image' version was not specified for the
promtail-debug docker image, resulting in a ~6 month old 'latest' image
as a build image. Updating this resulted in one more change: the newer
version has a different location for the 'dlv' binary.

The promtail-debug image would also hang waiting for a connection when
ran, this adds the same flag, `--continue`, as added to the loki debug
file in #1372.

**Special notes for your reviewer**:
I was interested in using the `-server.profiling_enabled=true` flag, to
enable pprof endpoints, and scrape with phlare. This flag is not
available in the non-debug version. Therefore I do not really care for
attaching a debugger right away. Open to removing that change though.

**Checklist**
- [x] Reviewed the `CONTRIBUTING.md` guide
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`