Backport 7804 into 2.7.1 (#7896)
use grpc for communicating with compactor for query time filtering of
data requested for deletion (#7804)

Manual backport of #7804 and #7814

**What this PR does / why we need it**:
Add grpc support to compactor for getting delete requests and gen number
for query time filtering.
Since these requests are internal to Loki, it would be good to use grpc
instead of HTTP same as all the internal requests we do in Loki.

I have added a new config for accepting the grpc address of the
compactor. I tried having just the existing config and detecting if it
is a grpc server, but it was hard to do it reliably, considering the
different deployment modes we support. I think it is safe to keep it the
same and eventually deprecate the existing config.

**Checklist**
- [x] Documentation added
- [x] Tests updated
- [x] `CHANGELOG.md` updated

(cherry picked from commit 1410808ee9f20917476fabaa78aa8849ba7c7d20)

Co-authored-by: Sandeep Sukhani <sandeep.d.sukhani@gmail.com>