Shut down query frontend gracefully
This PR fixes a bug in the shutdown process of the query frontend. In
order to shut down the frontend gracefully without cancelling any
in-flight operations/sub-requests, the frontend service has to wait
until all in-flight operations are finished. Only then its sub-services
may be stopped.

In order to control the deterministic order of the shutdown of the
service and its child services, child services need to be started with
their own context (not with their parent context). Otherwise, if the
context of the frontend is cancelled, also the contexts of the child
services are cancelled, leading to a parallel stopping of the parent and
its child services.

During the time of the shutdown, the `RoundTripGRCP` function still
needs to accept any round-tripping requests. This is, because the
frontend round tripper is wrapped into a query range tripper ware, which
splits incoming (HTTP) query requests into multiple (gRPC) query
requests if needed.

Signed-off-by: Danny Kopping <danny.kopping@grafana.com>
Signed-off-by: Christian Haudum <christian.haudum@gmail.com>