Revert "Update Ingester Rate Calculations (#7652)" (#7991)
This change was supposed to smooth out the per-second rate of streams so
distributors could react more uniformly to streams with variable rates.
In practice, it resulted in drastically more resource usage on
distributors because `GetStreamRates` now responds with all streams on
an ingester rather than only the streams seen in the last second. This
change also didn't seem to affect sharding precision one way or another.

If we'd still like to implement exponential smoothing for stream rates,
it should be implemented in distributors to keep communication overhead
to a minimum.