store: write overlapping chunks to multiple stores (#7988)
Signed-off-by: Ashwanth Goli <iamashwanth@gmail.com>

**What this PR does / why we need it**:
When a chunk falls on the period boundary it should be written to both
the stores as it gets referred to by the indexes on both sides. But we
currently skip the writes for subsequent periods if the chunk is found
in the cache. This PR ensures that overlapping chunks get written to
both the stores

**Which issue(s) this PR fixes**:
Fixes #7847 

**Special notes for your reviewer**:
Downside of this approach is that we end up writing the overlapping
chunk multiple times, once from each instance in the replica set. But
these duplicate writes would be infrequent given this occurs only on
schema changes

**Checklist**
- [X] Reviewed the `CONTRIBUTING.md` guide
- [X] Tests updated
- [x] `CHANGELOG.md` updated

Signed-off-by: Ashwanth Goli <iamashwanth@gmail.com>