fix bugs in logs results caching and its tests (#7925)
**What this PR does / why we need it**:
When a logs query results in an empty response, we cache it to avoid
doing that query again and respond straight away with an empty response.
However, we cache a single entry per time split interval with the query
itself to keep things simple. For example, if the time split config for
tenant `A` is `30m`, then queries for intervals `10m`-`20m` and
`21m`-`25m` would have the same cache key.

Here is roughly how cache hit is handled:

* If the new query is within the cached query bounds, return empty
results
* If the start of new query is before the start time of the cached
query, do a query from `newQuery.Start` to `cachedQuery.Start`
* If the response of last query is also empty, set `cachedQuery.Start` =
`newQuery.Start`
* If the end of new query is after the end time of the cached query, do
a query from `cachedQuery.End` to `newQuery.Start`
* If the response of last query is also empty, set `cachedQuery.End` =
`newQuery.End`
* If we have changes in `cachedQuery.Start/End`, update it in the cache.

The problem here is when we do queries to fill the gap, we sometimes do
queries for the range outside of what the user requested and respond
back without reducing the response to what the user requested. For
example, if the cached query is from `21m`-`25m` and the user query is
from `10m`-`15m`, we will query for the whole gap i.e `10m`-`21m`. If
there are logs from `15m`-`21m` in the response, we will unexpectedly
send it back to the user.

This PR takes care of this issue by extracting the data and sending back
only the user's requested logs.

I have also found the tests for logs results cache were incorrect. They
heavily use
[mergeLokiResponse](https://github.com/grafana/loki/blob/e2842c69c5e089389b697faa792e368f42f84c19/pkg/querier/queryrange/codec.go#L1014)
for building the test data, but they were not setting the `limit` in
test queries which was causing `mergeLokiResponse` to send back empty
results. This means were were always comparing two empty results, which
would always be the same.
This PR also takes care of fixing it, and relevant changes to correct
the test.

**Checklist**
- [x] Tests updated
- [x] `CHANGELOG.md` updated