Add drone env vars to test output to distinguish the 2 unit test runs. (#8093)
**What this PR does / why we need it**:

Add the drone step name and branch to the test output. These are used to
distinguish between the two unit tests runs in a drone build. One unit
test run is for the current branch, the other is for the branch to be
merged to. This change helps in not reporting the unit tests results
twice.

There is also a fix for the TestPromtail test. The missing newline made
the test reporting fail.

Signed-off-by: Michel Hollands <michel.hollands@grafana.com>