Change delimiter in  test  "sed" command to ':' (#8108)
The current delimiter is `/` which breaks when a branch name contains
that character. `:` is disallowed in branch names so should serve as a
suitable delimiter

This is the same as https://github.com/grafana/loki/pull/8107 but the
branch name doesn't include a `/`. That shouldn't be an issue after this
is merged