Promtail Docs: Clarifiy labels and timestamps set by cri stage (#7521)
Add description that `cri` stage not only adds labels for timestamp, but
that it also sets the timestamp of the log-entry before sending it.

Honestly I don't know if this is the intention or not, but this is how
things are implemented today and it makes the most sense. Even if not,
changing the behavior now would not be possible without breaking old
implicit assumptions, so better have more explicit documentation.

Also clarify that the cri-`flags` get extracted as a label.