skip compacting tables that do not fit in start and end date of desired schema config (#8127)
Before compacting tables, we look up the schema config for them. If we
do not find a schema config for a table, we skip them. However, the
schema lookup code was considering the start time of the schema and not
the end time.
This can cause the compactor to try compacting tables with the same
prefix as one of the previous schemas, but the table interval falls into
one of the subsequent schemas with a different config.

This PR fixes the issue and simplifies the code by reusing existing
schema lookup code.
I have also updated test cases to verify the fix works.