Alerts: Exclude '/schedulerpb.SchedulerForQuerier/QuerierLoop' route in LokiRequestLatency alert (#8138)
Excludes the route `/schedulerpb.SchedulerForQuerier/QuerierLoop` from
the `LokiRequestLatency` alert query config. We have found that this
route often has extremely high request latencies which will trigger the
alert constantly for us. Looking through the source code it seems like
it is expected to have high request latencies for this route and so in
our configuration have excluded the route. We can see that this route is
also being excluded in the [respective mimir
config](https://github.com/grafana/mimir/blob/main/operations/mimir-mixin-compiled/alerts.yaml#L33)
so believe it might be relevant to exclude here in loki as well