[helm] allow provisioned tokens to be created in different namespaces (#8081)
**What this PR does / why we need it**:

_Note_: This PR bumps the chart to 4.0, so it includes a few other
breaking changes we've been waiting to make.

The new helm chart has a few different ways to create tokens when
running in enterprise mode:

* The tokengen job creates the initial admin token
* The provisioner job creates the self-monitoring tenant if enabled
* The provisioner job can also create additional tenants if configured

The problem with these jobs is they put the generated read/write tokens
for these tenants in secrets in the same namespace as the loki
deploiyment. This is problematic if you need these tokens for a specific
application, for example a Grafana datasource, that is deployed in a
separate namespace.

This PR introduces two breaking changes, one of which is breaking as it
significantly changes how both the self monitoring and additional
tenants are configured.

First it adds the value `enterprise.adminToken.additionalNamespaces` and
moves `enterprise.adminTokenSecret` to `enterprise.adminToken.secret`.
If `additionalNamespaces` are included, additional secrets containing
the GEL admin token are created in those namespaces as well as the
release namespace.

Second, it changes the structure of defining a tenant from just a string
name, to an object containing a `name` and `secretNamespace` field. The
`secretNamespace` is the namespace the secret containing that tenants
token(s) will be placed in. In the case of the self-monitoring token, a
secret will still be created in the default namespace in addition to the
`secretNamespace` specified as that token is required by multiple
components.