ingester: Add `ingester_memory_streams_labels_bytes` metric (#8212)
Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>

This is to give visibility on `labels` size of in-memory streams on the
ingesters.

It caculate the size of labels via `len(labels.String())`. Meaning if
you have
```go
	lbs := labels.Labels{
		{Name: "foo", Value: "bar"},
		{Name: "instance", Value: "loki"},
	}
```
size would be number of bytes in `{foo="bar", instance="loki"}` which is
`28`. NOTE: `len(string)` gives [bytes length in
Go](https://pkg.go.dev/builtin#len)