stores/series: always return entries to pool
Instead of doing the `Get` inside a function and `Put` outside, do them
both outside. This makes it clearer that it's ok to `defer` before
checking for errors, which avoids dropping a pooled slice as garbage on
any error.

A particular case is `CardinalityExceededError` where the dropped slice
will be large, and which is treated as benign unless it happens on
all matchers.

Change the pooled objects to pointer type to avoid the Go runtime
creating a pointer wrapper, and we can stop silencing the warning about
this from the staticcheck linter.