Attempt to fairly benchmark key-value splitting
Three apporoaches were identified - so trying each one to see which is
the best performant.

Note that the TrimLeft function will potentially result in whitespace
that appeared in the value being stripped - which the other two
approaches do not do, however, the message format is relatively
under-specified, so this may not be a problem in practice.