Loki cloud integration instructions (and necessary mixin changes) (#8492)
**What this PR does / why we need it**:

This PR is a first stab at documentation for setting up the Grafana
Cloud Loki Integration to monitor a local Loki cluster, installed using
the helm chart, using Grafana Cloud.

In addition to instructions on how to collect the necessary kubernetes
metrics, there were also some things with the Loki mixin (which provides
dashboards, alerts, and rules to the integration) that needed to be
fixed. For example, our dashboards rely on there being a cluster label
on our recording rules, yet our mixin does not include this? Also, we
are still including `cortex-gw` panels even when the mixin is compile
with `internal_components: false`.

**Special notes for your reviewer**:

The changes to the mixin will not be reflected in the current version of
the integration. We need to merge this PR first then update the
integration.

**Checklist**
- [ ] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`

---------

Co-authored-by: Michel Hollands <42814411+MichelHollands@users.noreply.github.com>
Co-authored-by: J Stickler <julie.stickler@grafana.com>