Loki Getting Started: Fix example query that was not working (#8668)
**What this PR does / why we need it**:

Previously example query was `evaluate-loki_flog_1` this was not the
label used in flog. This results in query not returning any logs which
beginners may be unsure of the cause of the issue. This might also be an
issue if flog was running in the background and logs are not displayed.

Changed to `evaluate-loki-flog-1`.


**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Special notes for your reviewer**:
Trivial fix to "Getting Started" documentation

**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`