Add initial implementation of per-query limits (#8727)
**What this PR does / why we need it**:
Sometimes we want to limit the impact of a single query by imposing
limits that are stricter than the current tenant limit. E.g. the maximum
query length could be seven days but based on the query or an admins
decision a query should just have a maximum length of one day. This is
where per-request limits come into play. They are passed via the
`X-Loki-Query-Limit` header and extracted into the requests context.

It is the responsibility of the operator or admin that the header is
valid.

**Which issue(s) this PR fixes**:
Fixes #8762

**Checklist**
- [ ] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [x] `CHANGELOG.md` updated
- [x] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`

---------

Signed-off-by: Callum Styan <callumstyan@gmail.com>
Co-authored-by: Karsten Jeschkies <karsten.jeschkies@grafana.com>