Ruler: remote rule evaluation (#8744)
**What this PR does / why we need it**:
Adds the ability to evaluate recording & alerting rules against a given
`query-frontend`, allowing these queries to be executed with all the
parallelisation & optimisation that regular adhoc queries have. This is
important because with `local` evaluation all queries are
single-threaded, and rules that evaluate a large range/volume of data
may timeout or OOM the `ruler` itself, leading to missed metrics or
alerts.

When `remote` evaluation mode is enabled, the `ruler` effectively just
becomes a gRPC client for the `query-frontend`, which will dramatically
improve the reliability of the `ruler` and also drastically reduce its
resource requirements.

**Which issue(s) this PR fixes**:
This PR implements the feature discussed in
https://github.com/grafana/loki/pull/8129 (**LID 0002: Remote Rule
Evaluation**).