Retention: add selector detail to docs (#8808)
**What this PR does / why we need it**:
This adds more detail around what can and cannot be used in a `selector`
for retention. It was unclear whether arbitrary LogQL could be used or
not; this makes it explicit.

Also added some additional matchers to the examples to show that it can
be combinatorial.

---------

Signed-off-by: Danny Kopping <danny.kopping@grafana.com>
Co-authored-by: J Stickler <julie.stickler@grafana.com>
(cherry picked from commit c15048b50bc11a96ac6ba393ddc9c62d8784e960)