Fix: Ensure query limits tripperware wraps frontend. (#8789)
**What this PR does / why we need it**:
The query frontend would not enforce the per-request query limits
because the injected tripperware was wrapped *after* the frontend has
been created. That is a bug. The order must be changed so that the
`queryrange.LimitsMiddleware` runs *after* the
`querylimits.limitsMiddleware`.

**Which issue(s) this PR fixes**:
Fixes #8790

**Checklist**
- [ ] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`