Canary: support filtering / parsing logs with LogQL
The canary logs can be shipped to Loki in various formats. This allows
users of the canary to parse out the canaries logs from any format e.g. nested
json.

The alternative is to do something different for the canary logs to
prevent the logs being wrapped / manipulated (e.g. write to Loki directly, or use a different log pipeline) which limits the benefits of the canary.