promtail: support configurable polling interval
Currently, Promtail always polls files every 250ms. This PR allows the
polling period to be customized, starting at a minimum polling period
and exponentially increasing the polling frequency to a bounded maximum
if the file has no changes.

When file changes are detected, the polling frequency resets to the
minimum.

The default behavior is the existing behavior; both the minimum and
maximum polling intervals are set to 250ms.