mixins: Normalize headless service name for query-frontend/scheduler (#8880)
**What this PR does / why we need it**:
Mixins in general are rather complex pieces of code to consume. When
trying to use the loki mixins it wasn't apparent at first that I could
use a headless service for DNS SRV discovery for both query_frontend and
query_scheduler, especially because for frontend the service is named
`query_frontend_headless_service` and for the scheduler, it's named
`query_scheduler_service_discovery`. Furthermore, the query_frontend
mixin provides a non-headless version of the service while
query_scheduler doesn't. This PR aims at normalizing this for end users,
both headless services are now named `query_frontend_headless_service`
and `query_scheduler_headless_service` and both also provide a
non-headless version of the service.