Canary: support filtering / parsing logs with LogQL (#8871)
**What this PR does / why we need it**:

The canary logs can be shipped to Loki in various formats. This allows
users of the canary to parse out the canaries logs from any format e.g.
nested json.

The alternative is to do something different for the canary logs to
prevent the logs being wrapped / manipulated (e.g. write to Loki
directly, or use a different log pipeline) which limits the benefits of
the canary.

I've confirmed this works locally.

**Which issue(s) this PR fixes**:
Fixes #7775

**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`