Promtail: Add a new target for the Azure Event Hub (#8787)
**What this PR does / why we need it**:

We want to allow receiving logs from the Azure Cloud. To solve this
problem, we need to be able to configure Promtail to consume logs from
the Azure Events Hub.

To achieve this goal, we need to add a new configuration option to the
scrape configuration in Promtail and implement the component that, given
this configuration, can connect to the Azure Event Hub, parse received
data, and prepare it to be stored in Loki.


**Which issue(s) this PR fixes**:
Fix [#8788](https://github.com/grafana/loki/issues/8788)

# **Special notes for your reviewer**:

With the current implementation, I would like to leverage Azure Events
Hub's compatibility with a Kafka protocol and reuse code for a Kafka
target.

### Why I decided to create a new target instead of just reusing the
Kafka target:
1. Configuration for Kafka in this scenario requires fewer parameters.
For example, version and auth should have some specific values.
2. Message handling is different for Azure Logs. In addition to what we
are doing for the Kafka target, we want to fix JSON and split incoming
messages into multiple log lines before propagating it further.

### From the implementation perspective, there are two goals:
- Extract interfaces that are not Kafka configuration specific from the
Kafka target module. Splitting Kafka config parsing and target creation.

- To make the message handler/parser injectable or configurable: we
would like to fix JSON and split a message into multiple messages.


**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [x] Documentation added
- [x] Tests updated
- [x] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`

---------

Co-authored-by: Pablo <2617411+thepalbi@users.noreply.github.com>
Co-authored-by: Paschalis Tsilias <tpaschalis@users.noreply.github.com>
Co-authored-by: J Stickler <julie.stickler@grafana.com>