Ruler: Implement consistent rule evaluation jitter (#8896)
**What this PR does / why we need it**:
This PR replaces the previous random jitter with a consistent jitter.
While both are random, having the random jitter be applied
_consistently_ is essential for evaluating rules on a predictable
cadence.

If a rule is supposed to evaluate every minute, whether it evaluates at
(e.g.) `01:00` or `01:03.234` is irrelevant because the evaluation
_instant_ is not adjusted, so it will produce the same result whether
run at `01:00` or `01:03.234`. However, if 1000 rules are set to
evaluate at `01:00`, this will create a resource contention issue.