azure: respect retry config before cancelling the context (#8732)
**What this PR does / why we need it**:

On GET blob operations a context timeout is set to `RequestTimeout`,
however this value is not the full timeout. This is only the first
timeout and afterwards there are retries. As a result, the retry
configuration is never used because the context is immediately
cancelled.

To fix we should set the context timeout to a value larger than the
Azure timeout after retries. Since the backoff is exponential and
randomized we need to take the value of `MaxRetryDelay` to set an upper
bound.