Bumps loki.mixins dependency on grafana/jsonnet-lib
Problem: as of this commit the loki.mixins have its dependency for
ksonnet-util pinned on a commit from Aug 2020. Because of this the loki
mixins haven't enjoyed the many contributions done to the
grafana/jsonnet-libs project. Namely without this update it's not possible to
generate the loki ConfigMap if you are not using Tanka

Signed-off-by: JoaoBraveCoding <jmarcal@redhat.com>