Log when returning query-time limit (#8938)
**What this PR does / why we need it**:
At https://github.com/grafana/loki/pull/8727 we introduced various
limits that can now be configured at query time. We always compare the
value of the limit configured at query time with the value set on the
overrides for the tenant or the default if not configured (aka
original); applying the most restrictive one.

If the most restrictive is the original value or the limit is not
configured at query-time, we print the following debug message:

https://github.com/grafana/loki/blob/9e1846c47a1f3685fc540b7d03d285c7530da223/pkg/util/querylimits/limiter.go#L43

It will be printed many times: every time the query is not configured at
query-time, as well as every time the original value is more
restrictive. Moreover, this log message lacks some useful information
such as the original value, the query-time value, the tenant ID and the
limit name.

This PR fixes this by removing the debug message above and printing a
new debug message only if we return the query-time limit. This new log
also prints the tenant ID, the limit name, as well as the original and
the query-time limit values.

**Which issue(s) this PR fixes**:
Fixes https://github.com/grafana/loki/issues/8932

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`