Enforce per tenant queue size
Prior to the changes in https://github.com/grafana/loki/pull/8752 the
max queue size per tenant was enforced by the size of the buffered channel
to which a request was enqueued.
However, since we have hierarchical queues, every subqueue has the same
channel capacity as the root (tenant) queue. Therefore the total queue
size per tenant needs to be tracked separately, so requests can be
rejected when the max queue size is reached.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>