tsdb: use unique uploader name while building tsdb index filename, same as boltdb-shipper (#8975)
**What this PR does / why we need it**:
While building file names from `boltdb-shipper` and `tsdb`, we add the
name of the uploader to it.
To avoid collisions in filenames when running multiple ingesters with
the same name, we also add a nanosecond precise timestamp to the
ingester name, which is persistent as the uploader name in the active
index directory.

This is not being done for the tsdb index built from ingesters which can
cause collisions in filenames when running multiple ingesters with the
same name. I have moved the code to manage the uploader name to the
index shipper config and used the same to share it with tsdb.