index-gateway: fix failure in initializing index gateway when boltdb-shipper is not being used (#8755)
**What this PR does / why we need it**:
When only `tsdb` index is being used in the schema without configuring
`boltdb-shipper` index store, we fail to initialize `index-gateway`
since it also tries to build `boltdb-shipper` index client. This
reference is used for serving `QueryIndex` calls which are for serving
index from boltdb files.

This PR fixes the issue by adding a check for the existence of
`boltdb-shipper` in the schema config before initializing an index
client. On the index gateway side, it adds a `failingIndexClient`, which
errors when someone tries to query it.

**Checklist**
- [ ] `CHANGELOG.md` updated

---------

Co-authored-by: Travis Patterson <masslessparticle@gmail.com>
Co-authored-by: Dylan Guedes <djmgguedes@gmail.com>
Co-authored-by: Travis Patterson <travis.patterson@grafana.com>
(cherry picked from commit 1fcefe8efaccc4175506d2a64f7d4dbe55a5e608)