Promtail: Prevent logging errors on normal shutdown
Since PR#4066 two messages are logged for the errors `ErrExpired` and `EOF`,
whereas nothing has been logged beforehand.

As `ErrExpired` is returned by the `JournalReader` after sending a message
to the `until` channel to signal a normal stop, this should not be logged
as an error.

Also preventing to log the same error twice (only differing `msg`).